refactor: PrezelButton 컴포넌트 구조 개선 및 내부 로직 최적화
